Keys, Arrows and Debug Controls
A key painter has been added, along with an arrow icon. The former allows for keyboard keys to be effectively stamped around using the painter, and the latter allows for buttons with arrows for the debug menu. In regards to the debug menu, snake movement controls have been added to the debug menu, along with a toggle button for the gameplay timer and to hide the debug button from the main menu (used for taking pictures of gameplay). The key control icon and movement control icon have been modified to take advantage of the new  KeyPainter class, and the movement control icon is now a subclass of the key control icon. Finally, the action queue for snakes have been modified to store the consumer most recently removed via the next action methods, so that it can be factored into whether a consumer is skipped. Now, an action can be skipped if it matches the most recently removed action (assuming the snake skips repeated actions). This makes sure snakes don't overshoot if the player is holding down a movement key.