Color selection box stuff
I've added ColorSelectionPainter, ColorSelectionIcon, and DefaultColorSelectionIcon. ColorSelectionPainter is used to paint a color selection box, ColorSelectionIcon is an abstract icon version of ColorSelectionPainter which allows for a color to be provided via the abstract getColor method, and DefaultColorSelectionIcon is the default ColorSelectionIcon implementation which stores the color to be returned by getColor.