You can now set the colors
I've added functionality to allow for the user to set the colors for the snake, apples, and the tile background. It will still require implementing a configuration file to allow these changes to persist after the game is closed, but this is one step closer to my vision of this game.