Added an arrow painter
An arrow painter was added to allow arrows of a given direction flag value to be painted. ArrowIcon and ArrowKeyControlIcon have been updated to use ArrowPainter to paint their arrows. Additionally, methods to check for opposing direction flags have been added to SnakeUtilities along with a method to check the directions to use for a triangle and a method to get a triangle as a Path2D object. The setter methods for KeyPainter now return a KeyPainter with the same optional configuration argument datatype as the calling KeyPainter (they still return the calling KeyPainter, but now they return it correctly). JPlayField now determines if its model would change by checking to see if the models are the same instance instead of checking to see if they are equivalent. The snake controls on the debug window for the main game now uses a box layout with fillers in between the controls instead of a group layout. Some documentation was reworked.