Repository now excludes irrelevant project file
The project file that stores the currently open files and bookmarks for the files will now be excluded from the repository since it's only really relevant for me. The file can safely be excluded from the repository since it stores which files I currently have open for the project and what I've bookmarked in the project.