Game overlays now use the component
For most of the instances where the game overlay is changed, the overlay component is now what is being used instead of its index in the gameOverlayLayers array. Some documentation was reworked, and the java.awt. prefix was removed from EventQueue in the main method since we are already importing all the classes from the java.awt package. Oh, and I finally got it to ignore the nbproject/private/private.xml file.