Revised Documentation+Removed SNAKE4J_CONFIG_FILE
The SNAKE4J_CONFIG_FILE value was removed from SnakeGame4J (it wasn't really used for anything and only appeared in a pre-release version). None of the constructors for private nested classes are public anymore, they are either private or package. The bean info for JPlayField now imports the java.awt library and BufferedImage instead of specifying their packages. Some documentation was revised, such as the documentation for the AbstractPlayFieldModel iterator and Snake iterator.