Map Entry Classes, Player Two Components, and SnakeCommand fields
The main addition for this commit are 3 classes relating to map entries. The first one is AbstractMapEntry, which can store a key and implements the equals and hashCode methods as defined by Map.Entry. The second one is MapBackedEntry, which uses a backing map's get and put methods to get and set its value. The third one is AbstractMapEntryIterator, which is an iterator to help with implementing map entry iterators when using the keys in the map to get an entry. These three classes may be moved to a different package later on during this branch's life.

The next improvement is that SnakeGame4J now has some (though likely not all) of the components relating to the second player, namely the labels stating the controls and the button for setting the second player's color. These are currently hidden for now, but can be made visible using the showP2Toggle check box. In addition, the startGameButton button has been renamed to configGameButton, since it does not actually start the game. The gameSettingsStartButton button is now the startGameButton, since it is the button that starts the game.

Third thing that was changed is that SnakeCommands can now have an integer representing their direction (previously, this was retrieved via the getDirectionOf method, which is now deprecated). This direction can be retrieved via invoking the getDirection method on a SnakeCommand. Additionally, the getAction method can be used to get an immutable SnakeActionCommand for a SnakeCommand without having to use the map returned by getCommandActionMap (this was previously accomplished via the now deprecated static getActionForCommand method). All uses of SnakeCommand.getDirectionOf and SnakeCommand.getActionForCommand have now been replaced by calling getDirection and getAction, respectively, on the SnakeCommand in question.

ArrowKeyControlIcon now actually uses its ArrowPainter. This was previously overlooked when ArrowPainter was being added.

The package info for the src/snake/game package has been finally added.

There were some updates to the comments.